+++ lib/features/profile/presentation/pages/profile_page_builder.dart
@@ -6,11 +6,15 @@ import '../../../auth/domain/entities/user.dart';
 import '../../../auth/domain/repositories/auth_repository.dart';
 import '../../../auth/presentation/cubits/auth_session_cubit.dart';
 import '../../../auth/presentation/cubits/logout_cubit.dart';
+import '../../../subscriptions/domain/repositories/subscriptions_repository.dart';
+import '../../../subscriptions/presentation/cubits/cancel_subscription_cubit.dart';
 import '../../domain/repositories/profile_parameters_repository.dart';
 import '../../domain/repositories/profile_repository.dart';
 import '../../domain/repositories/profile_statistics_repository.dart';
 import '../cubits/delete_profile_cubit.dart';
 import '../cubits/profile_parameters_cubit.dart';
+import '../cubits/profile_refresh_cubit.dart';
+import '../cubits/profile_subscription_cubit.dart';
 import '../cubits/profile_statistics_cubit.dart';
 import '../cubits/profile_user_cubit.dart';
 import 'profile_page.dart';
@@ -47,10 +51,21 @@ class ProfilePageBuilder extends StatelessWidget {
           )..loadInitial(),
         ),
         BlocProvider(
+          create: (_) => ProfileSubscriptionCubit(
+            di<SubscriptionsRepository>(),
+          ),
+        ),
+        BlocProvider.value(
+          value: di<ProfileRefreshCubit>(),
+        ),
+        BlocProvider(
           create: (_) => LogoutCubit(di<AuthRepository>()),
         ),
         BlocProvider(
           create: (_) => DeleteProfileCubit(di<ProfileRepository>()),
+        ),
+        BlocProvider(
+          create: (_) => CancelSubscriptionCubit(di<SubscriptionsRepository>()),
         ),
       ],
       child: const ProfilePage(),