+++ lib/features/subscriptions/presentation/pages/subscriptions_details_page.dart
@@ -1,9 +1,12 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../../core/constants/app_assets.dart';
 import '../../../../core/constants/app_strings.dart';
+import '../../../../core/di/di.dart';
 import '../../../../core/router/router_paths.dart';
 import '../../../../uikit/buttons/app_back_button.dart';
 import '../../../../uikit/buttons/main_button.dart';
@@ -11,6 +14,7 @@ import '../../../../uikit/cards/app_card.dart';
 import '../../../../uikit/images/svg_picture_widget.dart';
 import '../../../../uikit/themes/colors/app_color_theme.dart';
 import '../../../../uikit/themes/text/app_text_theme.dart';
+import '../../../profile/presentation/cubits/profile_refresh_cubit.dart';
 import '../../domain/entities/subscription_catalog_item.dart';
 import '../cubits/subscription_details_cubit.dart';
 import '../cubits/subscription_payment_cubit.dart';
@@ -46,6 +50,7 @@ class SubscriptionsDetailsPage extends StatelessWidget {
       paymentCubit: context.read<SubscriptionPaymentCubit>(),
     );
     if (!context.mounted || didPay != true) return;
+    di<ProfileRefreshCubit>().requestRefresh();
     context.go(AppRoutePaths.profilePath);
   }